feat: add file asset support (#106)
Fixes #58 

Mimics file asset support from the Service Catalog L2 construct. Uses the S3 bucket deployment construct called in the parent stack to send artifacts to an S3 bucket passed a prop to the StackSet stack. The bucket deployment and resources needed to faciliate it are siblings to the `AWS::CloudFormation::StackSet` resource in the synthesized CloudFormation template. The asset bucket must be outside the CDK-owned asset bucket process and is passed as a prop.

We also need to ensure the bucket deployment process finishes before we try to run the StackSet so there is a dependency added after the creation of the StackSet resource. This is done using a construct name shared across the synthesis and StackSet creation processes. I'm open to feedback on if there is a better way to do this. But we do need to ensure the bucket deployment finishes before CloudFormation creates/updates the StackSet resource.

There is also potential to create another type of `StackSynthesizer` and let StackSet and Service Catalog share it since this logic is very similar to how service catalog does it.

I also plan to add some additional integration tests to validate more features of the file asset process but wanted to open the pull request to get eyes and feedback on the approach while finishing up the tests. I also think more scenarios could be added to the README to clear up any complexity.